requirements.txt
flask>=2.3.3
matplotlib>=3.8.0
seaborn>=0.13.0
flask-sqlalchemy>=3.1.1
flask-login>=0.6.3
werkzeug>=3.0.1
supabase>=2.3.0
python-dotenv>=1.0.0 